+++ libs/model/src/thread/index.ts
@@ -2,4 +2,5 @@ export * from './CreateThread.command';
 export * from './CreateThreadReaction.command';
 export * from './DeleteThread.command';
 export * from './GetThreads.query';
+export * from './SearchThreads.query';
 export * from './UpdateThread.command';
