+++ packages/commonwealth/server/controllers/server_threads_controller.ts
@@ -24,11 +24,6 @@ import {
   GetThreadsByIdResult,
   __getThreadsById,
 } from './server_threads_methods/get_threads_by_id';
-import {
-  SearchThreadsOptions,
-  SearchThreadsResult,
-  __searchThreads,
-} from './server_threads_methods/search_threads';
 
 /**
  * Implements methods related to threads
@@ -48,13 +43,6 @@ export class ServerThreadsController {
     options: GetActiveThreadsOptions,
   ): Promise<GetActiveThreadsResult> {
     return __getActiveThreads.call(this, options);
-  }
-
-  async searchThreads(
-    this: ServerThreadsController,
-    options: SearchThreadsOptions,
-  ): Promise<SearchThreadsResult> {
-    return __searchThreads.call(this, options);
   }
 
   async countThreads(
