+++ packages/commonwealth/server/routes/threads/get_threads_handler.ts
@@ -2,6 +2,7 @@ import { AppError, query } from '@hicommonwealth/core';
 import { Thread } from '@hicommonwealth/model';
 import * as schemas from '@hicommonwealth/schemas';
 import { GetThreadsOrderBy, GetThreadsStatus } from '@hicommonwealth/schemas';
+import { SearchThreads } from 'node_modules/@hicommonwealth/model/src/thread';
 import { z } from 'zod';
 import { ALL_COMMUNITIES } from '../../middleware/databaseValidationService';
 import { ServerControllers } from '../../routing/router';
@@ -36,6 +37,7 @@ export type ActiveThreadsRequestQuery = {
 export type SearchThreadsRequestQuery = {
   search: string;
   thread_title_only?: string;
+  order_by?: 'last_active' | 'rank' | 'created_at' | 'profile_name';
 } & PaginationQueryParams;
 export type BulkThreadsRequestQuery = {
   topic_id: string;
@@ -165,15 +167,18 @@ export const getThreadsHandler = async (
       throw new AppError(Errors.NoCommunity);
     }
 
-    const searchResults = await controllers.threads.searchThreads({
-      communityId: community_id,
-      searchTerm: search,
-      threadTitleOnly: thread_title_only === 'true',
-      limit: parseInt(limit!, 10) || 0,
-      page: parseInt(page!, 10) || 0,
-      orderBy: order_by,
-      orderDirection: order_direction as any,
-      includeCount: include_count,
+    const searchResults = await query(SearchThreads(), {
+      actor: { user: { email: '' } },
+      payload: {
+        communityId: community_id,
+        searchTerm: search,
+        threadTitleOnly: thread_title_only === 'true',
+        limit: parseInt(limit!, 10) || 0,
+        page: parseInt(page!, 10) || 0,
+        orderBy: order_by,
+        orderDirection: order_direction as any,
+        includeCount: include_count,
+      },
     });
     return success(res, searchResults);
   }
